requirements.txt
python-telegram-bot>=10.1.0
bs4
lxml
requests
uuid
